// src/integrations/supabase/types_generated/tables.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
      clients: {
        Row: {
          address: string | null
          cnpj: string | null
          contact_person: string | null
          created_at: string | null
          email: string | null
          id: string
          name: string
          phone: string | null
          user_id: string
        }
        Insert: {
          address?: string | null
          cnpj?: string | null
          contact_person?: string | null
          created_at?: string | null
          email?: string | null
          id?: string
          name: string
          phone?: string | null
          user_id: string
        }
        Update: {
          address?: string | null
          cnpj?: string | null
          contact_person?: string | null
          created_at?: string | null
          email?: string | null
          id?: string
          name?: string
          phone?: string | null
          user_id?: string
        }
        Relationships: []
      }
      coletas: {
        Row: {
          bairro: string | null
          cep: string | null
          cidade: string | null
          client_id: string | null
          cnpj: string | null
          contato: string | null
          contrato: string | null
          created_at: string | null
          email: string | null
          endereco: string | null
          id: string
          localidade: string | null
          modelo_aparelho: string | null
          nf_glbl: string | null
          nf_metodo: string | null
          observacao: string | null
          parceiro: string | null
          previsao_coleta: string | null
          qtd_aparelhos_solicitado: number | null
          responsavel: string | null
          responsible_user_id: string | null
          status_coleta: string | null
          status_unidade: string | null
          telefone: string | null
          type: string
          uf: string | null
          user_id: string
          // NEWLY ADDED COLUMNS
          cep_origem: string | null
          cep_destino: string | null
          endereco_origem: string | null
          endereco_destino: string | null
        }
        Insert: {
          bairro?: string | null
          cep?: string | null
          cidade?: string | null
          client_id?: string | null
          cnpj?: string | null
          contato?: string | null
          contrato?: string | null
          created_at?: string | null
          email?: string | null
          endereco?: string | null
          id?: string
          localidade?: string | null
          modelo_aparelho?: string | null
          nf_glbl?: string | null
          nf_metodo?: string | null
          observacao?: string | null
          parceiro?: string | null
          previsao_coleta?: string | null
          qtd_aparelhos_solicitado?: number | null
          responsavel?: string | null
          responsible_user_id?: string | null
          status_coleta?: string | null
          status_unidade?: string | null
          telefone?: string | null
          type?: string
          uf?: string | null
          user_id: string
          // NEWLY ADDED COLUMNS
          cep_origem?: string | null
          cep_destino?: string | null
          endereco_origem?: string | null
          endereco_destino?: string | null
        }
        Update: {
          bairro?: string | null
          cep?: string | null
          cidade?: string | null
          client_id?: string | null
          cnpj?: string | null
          contato?: string | null
          contrato?: string | null
          created_at?: string | null
          email?: string | null
          endereco?: string | null
          id?: string
          localidade?: string | null
          modelo_aparelho?: string | null
          nf_glbl?: string | null
          nf_metodo?: string | null
          observacao?: string | null
          parceiro?: string | null
          previsao_coleta?: string | null
          qtd_aparelhos_solicitado?: number | null
          responsavel?: string | null
          responsible_user_id?: string | null
          status_coleta?: string | null
          status_unidade?: string | null
          telefone?: string | null
          type?: string
          uf?: string | null
          user_id?: string
          // NEWLY ADDED COLUMNS
          cep_origem?: string | null
          cep_destino?: string | null
          endereco_origem?: string | null
          endereco_destino?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "coletas_client_id_fkey"
            columns: ["client_id"]
            isOneToOne: false
            referencedRelation: "clients"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "coletas_responsible_user_id_fkey"
            columns: ["responsible_user_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
      Coletas: {
        Row: {
          created_at: string
          id: number
        }
        Insert: {
          created_at?: string
          id?: number
        }
        Update: {
          created_at?: string
          id?: number
        }
        Relationships: []
      }
      drivers: {
        Row: {
          created_at: string | null
          id: string
          license_plate: string | null
          name: string
          phone: string | null
          status: string
          user_id: string
          vehicle_type: string | null
        }
        Insert: {
          created_at?: string | null
          id?: string
          license_plate?: string | null
          name: string
          phone?: string | null
          status?: string
          user_id: string
          vehicle_type?: string | null
        }
        Update: {
          created_at?: string | null
          id?: string
          license_plate?: string | null
          name?: string
          phone?: string | null
          status?: string
          user_id?: string
          vehicle_type?: string | null
        }
        Relationships: []
      }
      items: {
        Row: {
          collection_id: string | null
          created_at: string | null
          description: string | null
          id: string
          image_url: string | null
          model: string | null
          name: string
          quantity: number
          status: string
          user_id: string
        }
        Insert: {
          collection_id?: string | null
          created_at?: string | null
          description?: string | null
          id?: string
          image_url?: string | null
          model?: string | null
          name: string
          quantity: number
          status?: string
          user_id: string
        }
        Update: {
          collection_id?: string | null
          created_at?: string | null
          description?: string | null
          id?: string
          image_url?: string | null
          model?: string | null
          name?: string
          quantity?: number
          status?: string
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "items_collection_id_fkey"
            columns: ["collection_id"]
            isOneToOne: false
            referencedRelation: "coletas"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      products: {
        Row: {
          code: string
          created_at: string | null
          description: string | null
          id: string
          image_url: string | null
          model: string | null
          serial_number: string | null
          user_id: string
        }
        Insert: {
          code: string
          created_at?: string | null
          description?: string | null
          id?: string
          image_url?: string | null
          model?: string | null
          serial_number?: string | null
          user_id: string
        }
        Update: {
          code?: string
          created_at?: string | null
          description?: string | null
          id?: string
          image_url?: string | null
          model?: string | null
          serial_number?: string | null
          user_id?: string
        }
        Relationships: []
      }
      profiles: {
        Row: {
          avatar_url: string | null
          first_name: string | null
          id: string
          last_name: string | null
          role: string
          updated_at: string | null
          phone_number: string | null
        }
        Insert: {
          avatar_url?: string | null
          first_name?: string | null
          id: string
          last_name?: string | null
          role?: string
          updated_at?: string | null
          phone_number?: string | null
        }
        Update: {
          avatar_url?: string | null
          first_name?: string | null
          id?: string
          last_name?: string | null
          role?: string
          updated_at?: string | null
          phone_number?: string | null
        }
        Relationships: []
      }
// ...
      transportadoras: {
        Row: {
          address: string | null
          cnpj: string | null
          contact_person: string | null
          created_at: string | null
          email: string | null
          id: string
          name: string
          phone: string | null
          user_id: string
        }
        Insert: {
          address?: string | null
          cnpj?: string | null
          contact_person?: string | null
          created_at?: string | null
          email?: string | null
          id?: string
          name: string
          phone?: string | null
          user_id: string
        }
        Update: {
          address?: string | null
          cnpj?: string | null
          contact_person?: string | null
          created_at?: string | null
          email?: string | null
          id?: string
          name?: string
          phone?: string | null
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "transportadoras_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never

export type TablesUpdate<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Update: infer U
    }
    ? U
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Update: infer U
      }
      ? U
      : never
    : never